pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "weixin-channel"
version = "0.1.0"
description = "Unofficial Python SDK for the Weixin ClawBot/iLink channel protocol."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "weixin-channel contributors" }]
dependencies = [
  "cryptography>=42",
  "httpx[socks]>=0.28",
  "pydantic>=2.12",
  "qrcode>=8.0",
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/krkawzq/weixin-channel-sdk"
Repository = "https://github.com/krkawzq/weixin-channel-sdk"
Issues = "https://github.com/krkawzq/weixin-channel-sdk/issues"

[project.optional-dependencies]
media = ["pillow>=10"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.11"]

[project.scripts]
weixin-channel = "weixin_channel.cli:main"
weixin-channel-login = "weixin_channel.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/weixin_channel"]

[tool.hatch.build.targets.sdist]
include = [
  "src/weixin_channel/**",
  "README.md",
  "LICENSE",
  "CHANGELOG.md",
  "docs/**",
  "examples/**",
  "pyproject.toml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
